create 6 package, mianly refactor visualizations package and create an interface for it.
