+++ app/src/main/java/chpc/database/Db.java
@@ -1,4 +1,4 @@
-package chpc.models;
+package chpc.database;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
