+++ app/src/main/java/chpc/visualizations/Line.java
@@ -1,6 +1,5 @@
-package chpc.views;
+package chpc.visualizations;
 
-import chpc.views.View;
 import org.jfree.chart.ChartFactory;
 import org.jfree.chart.JFreeChart;
 import org.jfree.chart.block.BlockBorder;
@@ -9,7 +8,7 @@ import org.jfree.chart.plot.PlotOrientation;
 import org.jfree.chart.renderer.category.LineAndShapeRenderer;
 import org.jfree.data.category.DefaultCategoryDataset;
 
-import chpc.models.NHPIRecord;
+import chpc.dataLoader.NHPIRecord;
 
 import java.awt.*;
 import java.util.Set;
