+++ app/src/main/java/chpc/visualizations/Scatter.java
@@ -1,35 +1,34 @@
-package chpc.views;
+package chpc.visualizations;
 
-import chpc.views.View;
 import org.jfree.chart.JFreeChart;
 import org.jfree.chart.axis.DateAxis;
 import org.jfree.chart.axis.NumberAxis;
 import org.jfree.chart.plot.XYPlot;
-import org.jfree.chart.renderer.xy.XYSplineRenderer;
+import org.jfree.chart.renderer.xy.XYItemRenderer;
+import org.jfree.chart.renderer.xy.XYLineAndShapeRenderer;
 import org.jfree.data.time.Month;
 import org.jfree.data.time.TimeSeries;
 import org.jfree.data.time.TimeSeriesCollection;
 
-import chpc.models.NHPIRecord;
+import chpc.dataLoader.NHPIRecord;
 
 import java.awt.*;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.*;
 
-public class Time extends View {
+public class Scatter extends View {
   TimeSeriesCollection dataset;
 
-  public Time() {
-    setTitle("TimeSeries Chart");
-    chartType = "Time";
+  public Scatter() {
+    setTitle("Scatter Chart");
+    chartType = "Scatter";
     createChart();
   }
 
   public void addDataset() {
 
     // set series
-
     dataset = new TimeSeriesCollection();
     ArrayList<TimeSeries> allSeries = new ArrayList<>();
 
@@ -41,15 +40,13 @@ public class Time extends View {
       String geo = entry.getKey();
       Set<NHPIRecord> records = entry.getValue();
       allSeries.add(new TimeSeries(geo));
-      for (var r : records) {
+      for (NHPIRecord r : records) {
         String date = r.getRefDate();
         try {
           Date myDate = standardDateFormat.parse(date);
-          // test output
           System.out.println("string: " + date.substring(0, 4) + " " + date.substring(5) + " " + r.getValue());
           double time = Double.parseDouble(date.substring(0, 4)) +
               Double.parseDouble(date.substring(5)) / 100;
-          // test output
           System.out.println("time: " + time);
           allSeries.get(i).add(new Month(myDate), r.getValue());
 
@@ -67,21 +64,26 @@ public class Time extends View {
 
   public JFreeChart plotView() {
     XYPlot plot = new XYPlot();
-    XYSplineRenderer splinerenderer1 = new XYSplineRenderer();
-    // XYSplineRenderer splinerenderer2 = new XYSplineRenderer();
+    XYItemRenderer itemrenderer1 = new XYLineAndShapeRenderer(false, true);
+    // XYItemRenderer itemrenderer2 = new XYLineAndShapeRenderer(false, true);
 
     plot.setDataset(0, dataset);
-    plot.setRenderer(0, splinerenderer1);
+    plot.setRenderer(0, itemrenderer1);
     DateAxis domainAxis = new DateAxis("Year");
     plot.setDomainAxis(domainAxis);
     plot.setRangeAxis(new NumberAxis("NHPI"));
 
+    // plot.setDataset(1, dataset2);
+    // plot.setRenderer(1, itemrenderer2);
+    // plot.setRangeAxis(1, new NumberAxis("US$"));
+
     plot.mapDatasetToRangeAxis(0, 0);// 1st dataset to 1st y-axis
     // plot.mapDatasetToRangeAxis(1, 1); // 2nd dataset to 2nd y-axis
 
-    JFreeChart chart = new JFreeChart("Time series: New housing price index (NHPI) ",
+    JFreeChart scatterChart = new JFreeChart("Scatter: New housing price index (NHPI)",
         new Font("Serif", Font.BOLD, 18), plot, true);
-    return chart;
+
+    return scatterChart;
   }
 
 }
